legacy-agent-publish/pyproject.toml
[project]
name = "azure-openai-chatbot"
version = "0.1.0"
description = "Chatbot application using Azure OpenAI in Foundry with EntraID authentication"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "streamlit>=1.40.0",
    "openai>=1.57.0",
    "azure-identity>=1.19.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "black>=24.10.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
